refactor: chapter2---1.mybatis的注解方式应用 chapter3---1.初始化chapter3工程 新增扩展资料

+++ chapter3/src/main/java/com/pers/gfc/chapter3/mapper/RoleMapper.java
@@ -7,7 +7,7 @@ import com.pers.gfc.chapter3.po.Role;
  * 注册当前接口为映射器.
  */
 public interface RoleMapper {
-    Role getRole(Long id);
+    Role getRole(Integer id);
 
     int deleteRole(Long id);
